ENUM Working

+++ PI_Start/PI_Start/lll.h
@@ -21,8 +21,8 @@ extern double *B;
  
 void initStructsLLL(int dim);
 void computeGSO(long** base);
-void shiftVector(double** base, int k, int kl);
-void lll(double** base, double delta, int kmax);
+void shiftVector(long** base, int k, int kl);
+void lll(long** base, double delta, int kmax);
  
 
 #endif /* defined(__PI_Start__lll__) */
